Hero
Separate Hero and SkillArcher class
