+++ Valiant Guardian - Demon Invasion/Assets/Scripts/Hero/Hero.cs
@@ -8,7 +8,8 @@ public class Hero : MonoBehaviour
     public float protectionRadius = 10f;
     //    public float nearRadius = 3f;
 
-    public float SkillCoolDown = 5f;
+    public float normSkillCoolDown = 5f;
+    public float ultiSkillCooldown = 20f;
     public int SkillBoardNumber;
 
     public GameObject Enemies;
@@ -26,14 +27,12 @@ public class Hero : MonoBehaviour
     public int useSkillHash = Animator.StringToHash("useSkill");
     public int idleStateHash = Animator.StringToHash("Base Layer.Archer-Idle-Anim");
 
-    public GameObject Projectiles;
-
     void Start()
     {
         ProjectilePosTr = transform.GetChild(0);
         anim = GetComponent<Animator>();
         //set skill cool down to skillboard
-        GameObject.Find("SkillPanel").transform.GetChild(SkillBoardNumber).GetComponent<SkillBtn>().setSkillCoolDown(SkillCoolDown);
+        GameObject.Find("SkillPanel").transform.GetChild(SkillBoardNumber).GetComponent<SkillBtn>().setNormSkillCoolDown(normSkillCoolDown);
         coolDown = idleTime;
     }
 
@@ -62,37 +61,5 @@ public class Hero : MonoBehaviour
         //default anim scale speed is 100
         anim.speed = speed / speedScaleRatio;
         idleTime = idleTime / (speed / speedScaleRatio);
-    }
-
-    public void heroSkill()
-    {
-        if (anim != null)
-            anim.SetTrigger(useSkillHash);
-        GameObject.Find("SkillPanel").transform.GetChild(SkillBoardNumber).GetComponent<SkillBtn>().resetCoolingDown();
-    }
-
-    public void skillArcher()
-    {
-        StartCoroutine(rapidArrow());
-    }
-
-    IEnumerator rapidArrow()
-    {
-        for (int i = 0; i < 3; i++)
-        {
-            //Set projectile object to look at enemy
-            //Projectile object is a child of hero object
-            Quaternion direction = Quaternion.LookRotation(Enemies.transform.position - ProjectilePosTr.position, ProjectilePosTr.TransformDirection(Vector3.up));
-            ProjectilePosTr.rotation = new Quaternion(0, 0, direction.z, direction.w);
-            spawnNormProjectile();
-            yield return new WaitForSeconds(0.05f);
-        }
-    }
-
-    public void spawnNormProjectile()
-    {
-        //spawning projectile
-        //called from animation
-        Instantiate(Projectiles, ProjectilePosTr.position, ProjectilePosTr.rotation);
     }
 }
