+++ Valiant Guardian - Demon Invasion/Assets/Scripts/Hero/SkillArcher.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6a4b3fce5e000ef4694b7f6e38393696
+timeCreated: 1436793167
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
