+++ Valiant Guardian - Demon Invasion/Assets/_Scenes/Survival.unity
@@ -465,7 +465,7 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1086712247}
+      - m_Target: {fileID: 1086712248}
         m_MethodName: heroSkill
         m_Mode: 1
         m_Arguments:
@@ -1145,35 +1145,12 @@ RectTransform:
 Transform:
   m_PrefabParentObject: {fileID: 458406, guid: 1464d07503ddcb64998f6fd341de8ba6, type: 2}
   m_PrefabInternal: {fileID: 2117005445}
---- !u!1 &1086712246 stripped
-GameObject:
-  m_PrefabParentObject: {fileID: 185694, guid: 9190dbb652121fd4b801a1a1a8853b5d, type: 2}
-  m_PrefabInternal: {fileID: 1504159585}
---- !u!114 &1086712247
+--- !u!114 &1086712248 stripped
 MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 1086712246}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 3032ccefadaa06e4d8ef1e0b714d9d0d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  DistanceFromHero: 0
-  protectionRadius: 15
-  SkillCoolDown: 5
-  SkillBoardNumber: 0
-  Enemies: {fileID: 0}
-  ProjectilePosTr: {fileID: 0}
-  anim: {fileID: 0}
-  idleTime: 1
-  coolDown: 0
-  heroSpeed: 100
-  isAttackingHash: 973584577
-  useSkillHash: 1576347820
-  idleStateHash: 1146448895
-  Projectiles: {fileID: 165164, guid: 358277bd7dddb0542840e1bf0194ae10, type: 2}
+  m_PrefabParentObject: {fileID: 11487806, guid: 9190dbb652121fd4b801a1a1a8853b5d,
+    type: 2}
+  m_PrefabInternal: {fileID: 1504159585}
+  m_Script: {fileID: 11500000, guid: 6a4b3fce5e000ef4694b7f6e38393696, type: 3}
 --- !u!1 &1112867035
 GameObject:
   m_ObjectHideFlags: 0
@@ -1582,10 +1559,6 @@ Prefab:
     - target: {fileID: 410682, guid: 9190dbb652121fd4b801a1a1a8853b5d, type: 2}
       propertyPath: m_RootOrder
       value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 11408174, guid: 9190dbb652121fd4b801a1a1a8853b5d, type: 2}
-      propertyPath: attackSpeed
-      value: 100
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 9190dbb652121fd4b801a1a1a8853b5d, type: 2}
